updated src/store/spots for improved readability and maintainability
