+++ frontend/src/components/SpotForms/Deletespot.jsx
@@ -1,0 +1,10 @@
+export default function DeleteSpot() {
+  return (
+    <div id="delete-spot-container">
+      <h1>Confirm Delete</h1>
+      <p>Are you sure you want to remove this spot?</p>
+
+      <button id="delete-spot-button">No (Keep Spot)</button>
+    </div>
+  );
+}
