+++ frontend/src/components/SpotsList/SpotsList.jsx
@@ -2,7 +2,7 @@ import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { Link } from "react-router-dom";
 import { GoStarFill } from "react-icons/go";
-import { fetchSpots } from "../../store/spots";
+import { fetchSpots } from "../../store/spots/thunks";
 import "./SpotsList.css";
 
 const SpotsList = () => {
