+++ frontend/src/store/spots/types.js
@@ -1,0 +1,7 @@
+// * Action Type Constants **********
+// Consistent and centralized way to identify the types of actions being dispatched in the app. Ensures reliability, scalability, and maintainability in codebase.
+export const LOAD_SPOTS = "/spots/LOAD_SPOTS";
+export const LOAD_SPOT = "/spots/LOAD_SPOT";
+export const CREATE_SPOT = "/spots/CREATE";
+export const UPDATE_SPOT = "/spots/UPDATE";
+export const DELETE_SPOT = "/spots/DELETE";
