+++ frontend/src/store/store.js
@@ -1,10 +1,11 @@
 import { createStore, applyMiddleware, compose, combineReducers } from "redux";
 import thunk from "redux-thunk";
 import sessionReducer from "./session";
-import spotsReducer from "./spots";
+import spotsReducer from "./spots/reducer";
 import reviewsReducer from "./reviews";
 
 // * 1 Create Root Reducer *********
+// Makes Redux applications scalable as the state grows in complexity.
 const rootReducer = combineReducers({
   session: sessionReducer,
   spots: spotsReducer,
@@ -28,3 +29,15 @@ const configureStore = (preloadedState) => {
 };
 
 export default configureStore;
+
+//# If we were using redux toolkit, we simply code with below:
+// import { configureStore } from "@reduxjs/toolkit";
+// import spotsReducer from "./spots/reducer";
+
+// const store = configureStore({
+//   reducer: {
+//     spots: spotsReducer,
+//   },
+// });
+
+// export default store;
